pyproject.toml
[tool.poetry]
name = "prodil"
version = "0.1.0"
description = ""
authors = ["ahmetveburak"]

[tool.poetry.dependencies]
python = "^3.8"
pymongo = "^3.10.1"
mongoengine = "^0.20.0"
pyrogram = {git = "https://github.com/pyrogram/pyrogram.git", rev = "asyncio"}
python-decouple = "^3.3"
tgcrypto = "^1.2.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = {version = "^19.10b0", allow-prereleases = true}
mypy = "^0.782"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
